Refactor code: rename DependencyInjector to DependencyInjection
