--- src/Configuration/DependencyInjector/index.ts
@@ -1,1 +1,0 @@
-export * from './DependencyInjector';
