+++ src/Modules/Station/Components/PlaylistTabs/PlaylistTabs.tsx
@@ -1,5 +1,5 @@
 import { BaseComponent } from 'BaseComponent';
-import { Inject } from 'Configuration/DependencyInjector';
+import { Inject } from 'Configuration/DependencyInjection';
 import { IApplicationState } from 'Configuration/Redux';
 import { Dispatch } from 'Configuration/Redux';
 import { FavoriteSong, PlaylistSong } from 'Models';
