+++ src/Services/Http/UserServices/UserServices.ts
@@ -1,4 +1,4 @@
-import { Service } from 'Configuration/DependencyInjector';
+import { Service } from 'Configuration/DependencyInjection';
 import {
   AccessToken,
   EmailForm,
